start-qemu-daemon: update to support booting libostree images
The default behaviour should have not changed and instead added new
environment variables to support libostree qemu boot.

The key difference between libostree boot and a "normal" boot
is that we need to pass "-bios" instead "-kernel", and we provide
a disk image to "-drive" when booting with libostree instead of an
single ext4 image.

[GDP-613] integrate meta-updater

Signed-off-by: Mirza Krak <mirza.krak@endian.se>